simplify workflow wrapper
